+++ Engine/Draw/Object/BaseDrawObject.h
@@ -36,14 +36,14 @@ namespace YGame
 		/// </summary>
 		/// <param name="shaderTag"> : シェーダータグ</param>
 		/// <param name="drawPriority"> : 描画優先度</param>
-		void Draw(const std::string& shaderTag, const uint16_t drawPriority);
+		void Draw(const std::string& shaderTag, const size_t drawPriority);
 
 		/// <summary>
 		/// 描画
 		/// </summary>
 		/// <param name="shaderTag"> : シェーダータグ</param>
 		/// <param name="drawPriority"> : 描画優先度</param>
-		void Draw(const std::string& shaderTag, const uint16_t drawPriority, BaseGraphic* pGraphic);
+		void Draw(const std::string& shaderTag, const size_t drawPriority, BaseGraphic* pGraphic);
 	
 	public:
